+++ packages/babel-preset-granite/tsdown.config.ts
@@ -1,8 +1,7 @@
 import { defineConfig } from 'tsdown';
 
-export default defineConfig([
-  {
-    entry: ['src/index.ts'],
-    format: ['cjs'],
-  },
-]);
+export default defineConfig({
+  entry: ['src/index.ts'],
+  format: ['cjs'],
+  dts: true,
+});