+++ packages/cli/scripts/run.js
@@ -1,2 +1,1 @@
- 
 require('../dist').initialize().catch(console.error);