+++ packages/react-native/src/router/Router.tsx
@@ -56,6 +56,7 @@ export interface InternalRouterProps {
   container: ComponentType<PropsWithChildren<InitialProps>>;
   initialProps: InitialProps;
   initialScheme: string;
+  setIosSwipeGestureEnabled?: ({ isEnabled }: { isEnabled: boolean }) => Promise<void> | void;
 }
 
 export type RouterProps = StackNavigatorProps & NavigationContainerProps;
@@ -131,6 +132,7 @@ export function Router({
   defaultScreenOption,
   screenContainer,
   // Public props (StackNavigator)
+  setIosSwipeGestureEnabled,
   ...navigationContainerProps
 }: InternalRouterProps & RouterProps): ReactElement {
   const initialRouteName = useInitialRouteName({ prefix, initialScheme });
@@ -163,7 +165,7 @@ export function Router({
 
   return (
     <NavigationContainer ref={navigationRef} {...navigationContainerProps} linking={linkingOptions}>
-      <CanGoBackGuard canGoBack={canGoBack} onBack={onBack}>
+      <CanGoBackGuard canGoBack={canGoBack} onBack={onBack} setIosSwipeGestureEnabled={setIosSwipeGestureEnabled}>
         <Container {...initialProps}>
           <StackNavigator.Navigator initialRouteName={initialRouteName} screenOptions={screenOptions}>
             {Screens}