+++ packages/babel-preset-granite/CHANGELOG.md
@@ -1,5 +1,23 @@
 # babel-preset-granite
 
+## 0.1.22
+
+### Patch Changes
+
+- 487600d: apply plugin-transform-flow-strip-types first
+
+## 0.1.21
+
+## 0.1.20
+
+### Patch Changes
+
+- 1df5883: update package.json meta to supports any moduleResolutions
+
+## 0.1.19
+
+## 0.1.18
+
 ## 0.1.17
 
 ## 0.1.16