+++ packages/plugin-hermes/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @granite-js/plugin-hermes
 
+## 0.1.22
+
+### Patch Changes
+
+- @granite-js/plugin-core@0.1.22
+- @granite-js/utils@0.1.22
+
+## 0.1.21
+
+### Patch Changes
+
+- @granite-js/plugin-core@0.1.21
+- @granite-js/utils@0.1.21
+
+## 0.1.20
+
+### Patch Changes
+
+- @granite-js/plugin-core@0.1.20
+- @granite-js/utils@0.1.20
+
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [e9a3daf]
+  - @granite-js/plugin-core@0.1.19
+  - @granite-js/utils@0.1.19
+
+## 0.1.18
+
+### Patch Changes
+
+- @granite-js/plugin-core@0.1.18
+- @granite-js/utils@0.1.18
+
 ## 0.1.17
 
 ### Patch Changes